requirements.txt
streamlit>=1.31.0
groq>=0.4.0
torch>=2.2.0
transformers>=4.37.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.18.0
captum>=0.6.0
shap>=0.44.0
scikit-learn>=1.3.0
networkx>=3.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=8.0.0
python-dotenv>=1.0.0 